refactor: remove unused code and update component package
